backend/requirements.txt
fastapi
uvicorn[standard]
python-multipart
pydantic
google-genai
PyMuPDF
python-dotenv